od: Fix for well import issues and wrong time-depth model
Summary: Simple vertical wells were not loading, and corrected a problem in PDO dataset.

Test Plan: added these two new cases to the library of auto-tested well imports (now with 17 items). All tests are successful.

Reviewers: nanne

Reviewed By: nanne

Subscribers: kristofer, mahant

Differential Revision: https://phabricator.dgbes.com:8443/D7326

git-svn-id: https://svn.opendtect.org/od/trunk@38598 a985855d-d384-4b45-ae33-31a410998c3c

+++ include/Well/wellimpasc.h
@@ -112,11 +112,11 @@ protected:
 
     od_istream&			strm_;
     bool			readTrackData(TypeSet<Coord3>&,
-					      TypeSet<float>&,
-					      float& kbelevinfile) const;
+					      TypeSet<double>&,
+					      double& kbelevinfile) const;
     bool			computeMissingValues(TypeSet<Coord3>&,
-						     TypeSet<float>&,
-						     float& kbelevinfile) const;
+						     TypeSet<double>&,
+						    double& kbelevinfile) const;
     bool			adjustSurfaceLocation(TypeSet<Coord3>&,
 						     Coord& surfacecoord) const;
